removed old query from sqlite functions not needed anymore
